Removed warnings
